+++ R/globals_extra.R
@@ -1,5 +1,5 @@
 utils::globalVariables(c(
   "frac_optimal", "frac_LL", "response_rate", "breakpoint", "frac_risky",
   "paradigm", "score", "reinforcement_rate", "delay", "v", "value", "name", "episode",
-  "step", "responses", "elapsed", "window_mid", "rate", "response"
+  "step", "responses", "elapsed", "window_mid", "rate", "response", "condition", "response_rate", "is_control"
 ))